refactor(types): migrate type declaration files from .d.tsx to .types.ts
- Rename all component type files from the `.d.tsx` convention to `.types.ts`
across Organisms and UI components, and update all import paths accordingly.
- Improve type specificity in ResourcesLink (typed array, optional download prop)
and Pagination (clarified JSDoc). Reformat chat-gpt HTML and update docs.
