+++ src/Components/Organisms/DictionarySearch/DictionarySearch.types.ts
