+++ src/Components/Organisms/Footer/ResourcesLink/ResourcesLink.types.ts
@@ -1,11 +1,11 @@
 /**
  * Defines the props for the ResourcesLink component.
- * @param linkData - The data for the link.
- * @param linkClasses - The CSS classes to apply to the link.
+ * @param linkData - An array of resource link data objects to render.
+ * @param linkClasses - The CSS classes to apply to each link.
  */
 
 export interface IResourceLinkProps {
-  linkData: any,
+  linkData: IResourceLinkDataProps[],
   linkClasses: string | undefined
 }
 
@@ -14,7 +14,7 @@ export interface IResourceLinkProps {
  * @param label - The label for the link.
  * @param href - The URL for the link.
  * @param target - The target for the link (e.g. "_blank" to open in a new tab).
- * @param download - Whether the link should trigger a file download.
+ * @param download - A string filename to trigger a file download, or false to disable it.
  * @param rel - The rel attribute for the link.
  * @param type - The type attribute for the link.
  * @param icon - The icon to display for the link.
@@ -24,7 +24,7 @@ export interface IResourceLinkDataProps {
   label: string;
   href: string;
   target: string;
-  download: boolean;
+  download?: string | boolean;
   rel: string;
   type: string;
   icon: string;
