+++ src/Components/Organisms/Pronounciation/Pronounciation.types.ts
