+++ src/Components/Organisms/Pronounciation/PronounciationList/PronounciationList.types.ts
