+++ src/Components/Organisms/TwoCol/TwoCol.types.ts
