+++ src/Components/Organisms/WordAttributes/WordAttributes.types.ts
