+++ src/Components/Organisms/WordOfTheDay/WordOfTheDay.types.ts
